fix(TDI-44088) jdbc test perf
