--- jdbc/src/test/java/org/talend/components/jdbc/suite/InputTest.java
@@ -1,149 +1,0 @@
-/*
- * Copyright (C) 2006-2020 Talend Inc. - www.talend.com
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package org.talend.components.jdbc.suite;
-
-import java.util.List;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInfo;
-import org.talend.components.jdbc.configuration.InputQueryConfig;
-import org.talend.components.jdbc.configuration.InputTableNameConfig;
-import org.talend.components.jdbc.dataset.SqlQueryDataset;
-import org.talend.components.jdbc.dataset.TableNameDataset;
-import org.talend.components.jdbc.datastore.JdbcConnection;
-import org.talend.components.jdbc.output.platforms.PlatformFactory;
-import org.talend.sdk.component.api.record.Record;
-import org.talend.sdk.component.runtime.manager.chain.Job;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.talend.sdk.component.junit.SimpleFactory.configurationByExample;
-
-@DisplayName("Input")
-public abstract class InputTest extends JDBCBaseTest {
-
-    @Test
-    @DisplayName("Query - valid select query")
-    void validQuery(final TestInfo testInfo) {
-        final int rowCount = 50;
-        final String testTableName = getTestTableName(testInfo);
-        insertRows(testTableName, rowCount, false, null);
-        final SqlQueryDataset sqlQueryDataset = new SqlQueryDataset();
-        final JdbcConnection connection = newConnection();
-        sqlQueryDataset.setConnection(connection);
-        sqlQueryDataset.setFetchSize(rowCount / 3);
-        sqlQueryDataset
-                .setSqlQuery("select * from " + PlatformFactory.get(connection, getI18nMessage()).identifier(testTableName));
-        final InputQueryConfig config = new InputQueryConfig();
-        config.setDataSet(sqlQueryDataset);
-        final String configURI = configurationByExample().forInstance(config).configured().toQueryString();
-        Job.components().component("jdbcInput", "Jdbc://QueryInput?" + configURI).component("collector", "test://collector")
-                .connections().from("jdbcInput").to("collector").build().run();
-
-        final List<Record> collectedData = getComponentsHandler().getCollectedData(Record.class);
-        assertEquals(rowCount, collectedData.size());
-    }
-
-    @Test
-    @DisplayName("Query - unvalid query ")
-    void invalidQuery(final TestInfo testInfo) {
-        final SqlQueryDataset sqlQueryDataset = new SqlQueryDataset();
-        final JdbcConnection connection = newConnection();
-        sqlQueryDataset.setConnection(connection);
-        sqlQueryDataset.setSqlQuery("select fromm " + getTestTableName(testInfo));
-        final InputQueryConfig config = new InputQueryConfig();
-        config.setDataSet(sqlQueryDataset);
-        final String configURI = configurationByExample().forInstance(config).configured().toQueryString();
-        assertThrows(IllegalStateException.class, () -> Job.components().component("jdbcInput", "Jdbc://QueryInput?" + configURI)
-                .component("collector", "test://collector").connections().from("jdbcInput").to("collector").build().run());
-    }
-
-    @Test
-    @DisplayName("Query -  non authorized query (drop table)")
-    void unauthorizedDropQuery() {
-        final SqlQueryDataset dataset = new SqlQueryDataset();
-        dataset.setConnection(newConnection());
-        dataset.setSqlQuery("drop table abc");
-        final InputQueryConfig config = new InputQueryConfig();
-        config.setDataSet(dataset);
-        final String configURI = configurationByExample().forInstance(config).configured().toQueryString();
-        assertThrows(IllegalArgumentException.class,
-                () -> Job.components().component("jdbcInput", "Jdbc://QueryInput?" + configURI)
-                        .component("collector", "test://collector").connections().from("jdbcInput").to("collector").build()
-                        .run());
-    }
-
-    @Test
-    @DisplayName("Query -  non authorized query (insert into)")
-    void unauthorizedInsertQuery() {
-        final SqlQueryDataset dataset = new SqlQueryDataset();
-        dataset.setConnection(newConnection());
-        dataset.setSqlQuery("INSERT INTO users(id, name) VALUES (1, 'user1')");
-        final InputQueryConfig config = new InputQueryConfig();
-        config.setDataSet(dataset);
-        final String configURI = configurationByExample().forInstance(config).configured().toQueryString();
-        assertThrows(IllegalArgumentException.class,
-                () -> Job.components().component("jdbcInput", "Jdbc://QueryInput?" + configURI)
-                        .component("collector", "test://collector").connections().from("jdbcInput").to("collector").build()
-                        .run());
-    }
-
-    @Test
-    @DisplayName("TableName - valid table name")
-    void validTableName(final TestInfo testInfo) {
-        final int rowCount = 50;
-        final String testTableName = getTestTableName(testInfo);
-        insertRows(testTableName, rowCount, false, null);
-        final InputTableNameConfig config = new InputTableNameConfig();
-        config.setDataSet(newTableNameDataset(testTableName));
-        final String configURI = configurationByExample().forInstance(config).configured().toQueryString();
-        Job.components().component("jdbcInput", "Jdbc://TableNameInput?" + configURI).component("collector", "test://collector")
-                .connections().from("jdbcInput").to("collector").build().run();
-
-        final List<Record> collectedData = getComponentsHandler().getCollectedData(Record.class);
-        assertEquals(rowCount, collectedData.size());
-    }
-
-    @Test
-    @DisplayName("TableName - invalid table name")
-    void invalidTableName() {
-        final TableNameDataset dataset = new TableNameDataset();
-        dataset.setConnection(newConnection());
-        dataset.setTableName("xxx");
-        final InputTableNameConfig config = new InputTableNameConfig();
-        config.setDataSet(dataset);
-        final String configURI = configurationByExample().forInstance(config).configured().toQueryString();
-        assertThrows(IllegalStateException.class,
-                () -> Job.components().component("jdbcInput", "Jdbc://TableNameInput?" + configURI)
-                        .component("collector", "test://collector").connections().from("jdbcInput").to("collector").build()
-                        .run());
-    }
-
-    @Test
-    @DisplayName("TableName - valid table name with null values")
-    void validTableNameWithNullValues(final TestInfo testInfo) {
-        final int rowCount = 1;
-        final String testTableName = getTestTableName(testInfo);
-        insertRows(testTableName, rowCount, true, null);
-        final InputTableNameConfig config = new InputTableNameConfig();
-        config.setDataSet(newTableNameDataset(testTableName));
-        final String configURI = configurationByExample().forInstance(config).configured().toQueryString();
-        Job.components().component("jdbcInput", "Jdbc://TableNameInput?" + configURI).component("collector", "test://collector")
-                .connections().from("jdbcInput").to("collector").build().run();
-
-        final List<Record> collectedData = getComponentsHandler().getCollectedData(Record.class);
-        assertEquals(rowCount, collectedData.size());
-    }
-
-}
