+++ jdbc/src/test/java/org/talend/components/jdbc/suite/containers/MariaContainerTest.java
@@ -10,20 +10,31 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.talend.components.jdbc.suite.uiaction;
+package org.talend.components.jdbc.suite.containers;
 
 import org.junit.jupiter.api.Tag;
 import org.talend.components.jdbc.containers.JdbcTestContainer;
 import org.talend.components.jdbc.containers.MariaDBTestContainer;
-import org.talend.components.jdbc.suite.UIActionServiceTest;
-import org.talend.sdk.component.junit5.WithComponents;
+import org.talend.components.jdbc.suite.JDBCBaseContainerTest;
 
 @Tag("IT")
-@WithComponents("org.talend.components.jdbc")
-public class MariaUIActionTest extends UIActionServiceTest {
+public class MariaContainerTest extends JDBCBaseContainerTest {
 
     @Override
     public JdbcTestContainer buildContainer() {
         return new MariaDBTestContainer();
     }
+
+    public class MariaPlatform extends PlatformTests {
+    }
+
+    public class MariaInput extends InputTest {
+    }
+
+    public class MariaOutput extends OutputTest {
+    }
+
+    public class MariaUI extends UIActionServiceTest {
+    }
+
 }
