+++ jdbc/src/test/java/org/talend/components/jdbc/suite/containers/MySQLContainerTest.java
@@ -10,20 +10,31 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.talend.components.jdbc.suite.output;
+package org.talend.components.jdbc.suite.containers;
 
 import org.junit.jupiter.api.Tag;
 import org.talend.components.jdbc.containers.JdbcTestContainer;
 import org.talend.components.jdbc.containers.MySQLTestContainer;
-import org.talend.components.jdbc.suite.OutputTest;
-import org.talend.sdk.component.junit5.WithComponents;
+import org.talend.components.jdbc.suite.JDBCBaseContainerTest;
 
 @Tag("IT")
-@WithComponents("org.talend.components.jdbc")
-public class MySQLOutputTest extends OutputTest {
+public class MySQLContainerTest extends JDBCBaseContainerTest {
 
     @Override
     public JdbcTestContainer buildContainer() {
         return new MySQLTestContainer();
     }
+
+    public class MySQLPlatform extends JDBCBaseContainerTest.PlatformTests {
+    }
+
+    public class MySQLInput extends JDBCBaseContainerTest.InputTest {
+    }
+
+    public class MySQLOutput extends JDBCBaseContainerTest.OutputTest {
+    }
+
+    public class MySQLUI extends JDBCBaseContainerTest.UIActionServiceTest {
+    }
+
 }
